refactor: :truck: move SatelliteObject interface to own file
